Unformatted/Unchanged new entries
I simply added the new entries. At the moment they are completly unformated and the keys still need to be changed. As my JabRef configuration is not set up for the structure that is currently used, we have to wait for the format update to fully include the bib. into the altruism project.